Magento Release 2.4.1-p1

+++ lib/internal/Magento/Framework/Filesystem/Directory/Write.php
@@ -8,6 +8,8 @@ namespace Magento\Framework\Filesystem\Directory;
 
 use Magento\Framework\Exception\FileSystemException;
 use Magento\Framework\Exception\ValidatorException;
+use Magento\Framework\Filesystem\DriverInterface;
+use Magento\Framework\Phrase;
 
 /**
  * Write Interface implementation
@@ -25,14 +27,14 @@ class Write extends Read implements WriteInterface
      * Constructor
      *
      * @param \Magento\Framework\Filesystem\File\WriteFactory $fileFactory
-     * @param \Magento\Framework\Filesystem\DriverInterface $driver
+     * @param DriverInterface $driver
      * @param string $path
      * @param int $createPermissions
      * @param PathValidatorInterface|null $pathValidator
      */
     public function __construct(
         \Magento\Framework\Filesystem\File\WriteFactory $fileFactory,
-        \Magento\Framework\Filesystem\DriverInterface $driver,
+        DriverInterface $driver,
         $path,
         $createPermissions = null,
         ?PathValidatorInterface $pathValidator = null
@@ -48,13 +50,13 @@ class Write extends Read implements WriteInterface
      *
      * @param string $path
      * @return void
-     * @throws \Magento\Framework\Exception\FileSystemException
+     * @throws FileSystemException|ValidatorException
      */
     protected function assertWritable($path)
     {
         if ($this->isWritable($path) === false) {
             $path = $this->getAbsolutePath($path);
-            throw new FileSystemException(new \Magento\Framework\Phrase('The path "%1" is not writable.', [$path]));
+            throw new FileSystemException(new Phrase('The path "%1" is not writable.', [$path]));
         }
     }
 
@@ -63,7 +65,7 @@ class Write extends Read implements WriteInterface
      *
      * @param string $path
      * @return void
-     * @throws \Magento\Framework\Exception\FileSystemException
+     * @throws FileSystemException
      */
     protected function assertIsFile($path)
     {
@@ -71,7 +73,7 @@ class Write extends Read implements WriteInterface
         clearstatcache(true, $absolutePath);
         if (!$this->driver->isFile($absolutePath)) {
             throw new FileSystemException(
-                new \Magento\Framework\Phrase('The "%1" file doesn\'t exist.', [$absolutePath])
+                new Phrase('The "%1" file doesn\'t exist.', [$absolutePath])
             );
         }
     }
@@ -149,7 +151,7 @@ class Write extends Read implements WriteInterface
      * @param string $destination
      * @param WriteInterface $targetDirectory [optional]
      * @return bool
-     * @throws \Magento\Framework\Exception\FileSystemException
+     * @throws FileSystemException
      * @throws ValidatorException
      */
     public function createSymlink($path, $destination, WriteInterface $targetDirectory = null)
@@ -178,10 +180,18 @@ class Write extends Read implements WriteInterface
     {
         $exceptionMessages = [];
         $this->validatePath($path);
+
         if (!$this->isExist($path)) {
             return true;
         }
+
         $absolutePath = $this->driver->getAbsolutePath($this->path, $path);
+        $basePath = $this->driver->getRealPathSafety($this->driver->getAbsolutePath($this->path, ''));
+
+        if ($path !== null && $path !== '' && $this->driver->getRealPathSafety($absolutePath) === $basePath) {
+            throw new FileSystemException(new Phrase('The path "%1" is not writable.', [$path]));
+        }
+
         if ($this->driver->isFile($absolutePath)) {
             $this->driver->deleteFile($absolutePath);
         } else {
@@ -198,12 +208,13 @@ class Write extends Read implements WriteInterface
 
             if (!empty($exceptionMessages)) {
                 throw new FileSystemException(
-                    new \Magento\Framework\Phrase(
+                    new Phrase(
                         \implode(' ', $exceptionMessages)
                     )
                 );
             }
         }
+
         return true;
     }
 
@@ -231,7 +242,7 @@ class Write extends Read implements WriteInterface
         }
         if (!empty($exceptionMessages)) {
             throw new FileSystemException(
-                new \Magento\Framework\Phrase(
+                new Phrase(
                     \implode(' ', $exceptionMessages)
                 )
             );
@@ -297,7 +308,7 @@ class Write extends Read implements WriteInterface
      *
      * @param string|null $path
      * @return bool
-     * @throws \Magento\Framework\Exception\FileSystemException
+     * @throws FileSystemException
      * @throws ValidatorException
      */
     public function isWritable($path = null)
@@ -313,7 +324,7 @@ class Write extends Read implements WriteInterface
      * @param string $path
      * @param string $mode
      * @return \Magento\Framework\Filesystem\File\WriteInterface
-     * @throws \Magento\Framework\Exception\FileSystemException
+     * @throws FileSystemException
      * @throws ValidatorException
      */
     public function openFile($path, $mode = 'w')
@@ -334,7 +345,7 @@ class Write extends Read implements WriteInterface
      * @param string $content
      * @param string|null $mode
      * @return int The number of bytes that were written.
-     * @throws FileSystemException
+     * @throws FileSystemException|ValidatorException
      */
     public function writeFile($path, $content, $mode = 'w+')
     {
@@ -344,7 +355,7 @@ class Write extends Read implements WriteInterface
     /**
      * Get driver
      *
-     * @return \Magento\Framework\Filesystem\DriverInterface
+     * @return DriverInterface
      */
     public function getDriver()
     {
